% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classy_acc.R
\name{classy_acc}
\alias{classy_acc}
\title{Classy Accuracy}
\usage{
classy_acc(pred, real)
}
\arguments{
\item{pred}{Vector of predicted outcomes}

\item{real}{Outcomes the algortihm is attempting to predict}
}
\description{
Quickly compute the accuracy of an algothim by indputing predicted and actual values of an outcome
}
\examples{


classy_acc(pred, real)

}
\keyword{classification}
\keyword{metric}
