% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{form.it}
\alias{form.it}
\title{formats a numeric vector to a fixed string format with digits}
\usage{
form.it(x, digits = 3, perc = F, max = NULL)
}
\description{
formats a numeric vector to a fixed string format with digits
}
