% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlations.R
\name{get_correlations}
\alias{get_correlations}
\title{Get Correlations}
\usage{
get_correlations(data, target)
}
\arguments{
\item{data}{Data Frame for which the correlations will be calculated}

\item{target}{OUtcome Vector to calculate pairwise correlations with}
}
\description{
Given a dataset, calculate the correlation coefficient of each variable with a
chosen target variable. The number of observations in the data frame must match
the length of the target vector. Returns a data frame listing the feature names
and associated correlation coefficient.
}
\examples{


get_correlations(data, data$target)

}
\keyword{Calculations}
\keyword{Correlation}
\keyword{Quick}
