% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_dst.R
\name{grab_dst}
\alias{grab_dst}
\title{raw data pull for DST data}
\usage{
grab_dst(path, reg, years, pop, vars = NULL, s.id = NULL, show.vars = F)
}
\arguments{
\item{path}{path to raw data dir}

\item{reg}{base name of registry (data file base word, when \link{0-9} are removed)}

\item{years}{a vector of the years (e.g. 1991:2021) or other regular definitions of time periods covered by registries. For DREAM, this would be CCYYMM-CCYYMM.}

\item{pop}{a vector of identifiers, which the data extracted will be limited to}

\item{vars}{vector of variable names to be kept}

\item{s.id}{name of the variable that contains the ids in source data}

\item{show.vars}{logical to enable printing of variable names in source data.}
}
\description{
raw data pull for DST data
}
