% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{is.ok.num}
\alias{is.ok.num}
\title{a set of generic functions that speed up coding}
\usage{
is.ok.num(x)
}
\description{
check string data before converting to numeric
}
\details{
returning true if character column can be converted to numeric without any losses
}
