% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcfb.R
\name{pcfb}
\alias{pcfb}
\title{Percent Change from Baseline}
\usage{
pcfb(data, var, groupvar = USUBJID, timevar = VISIT, bltime = "Baseline",
  reftime = "Week 2")
}
\arguments{
\item{data}{Dataset}

\item{var}{Variable of Interest}

\item{groupvar}{Variable to compute change within}

\item{timevar}{Variable that records timepoint of measurement}

\item{bltime}{Value of the baseline time}

\item{reftime}{Value of the reference time}
}
\description{
Calculates the percent % change in value of a variable over time - typically a baseline value.
Defaults references are to a common CDISC complient data structure.
}
\examples{
cfb(data = data, var = AVAL, groupvar = USUBJID,
timevar = VISIT)

}
\keyword{wrangling}
