% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clinical_data.R
\name{simulate_clinical_data}
\alias{simulate_clinical_data}
\title{Simulate Clinical Data}
\usage{
simulate_clinical_data(n, t)
}
\arguments{
\item{n}{Number of subjects for the simulated data}

\item{t}{Number of timepoints for each subject}
}
\description{
Simulates a simple dataset based on two parameters; the number of subjects and timepoints.
}
\examples{
simulate_sample_data(n = 10, t = 6)
}
\keyword{cats}
