% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgears.R
\name{skew.ci}
\alias{skew.ci}
\title{calculate ci for skewed distribution point estimates}
\usage{
skew.ci(x, rep = 1000, stat = "median", alt = "two.sided", alpha = 0.95)
}
\arguments{
\item{x}{variable to derive ci from}

\item{rep}{number of bootstrap draws}

\item{stat}{point estimate statistic to calculate ci for}

\item{alt}{one or two-sided testing}

\item{alpha}{alpha level}
}
\value{
returns a vector of length 2 with lower and upper interval limits
}
\description{
calculate ci for skewed distribution point estimates
}
