% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_n.R
\name{twoway.n}
\alias{twoway.n}
\title{twoway table of n indicator data}
\usage{
twoway.n(data, x, group)
}
\arguments{
\item{data}{data frame with the group identifier and n-counter}

\item{x}{the n count, should be just a konstant k=1, unless data is aggregated}

\item{group}{the name of the variable which identifies groups/arms.}
}
\description{
twoway table of n indicator data
}
