% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_chi.R
\name{twoway_chi}
\alias{twoway_chi}
\title{twoway table of categorical data}
\usage{
twoway_chi(data, x, group, bin = F, cens, force.two = F, show.na = F)
}
\arguments{
\item{data}{dataset containing group factor and the x-variable to be summed up}

\item{x}{variable to be summed}

\item{bin}{logical indicator to only use last row of two-way table (FALSE is default)}

\item{cens}{positive integer deciding the lower threshold for when to clear cells in order to hide very detailed patient-level data}

\item{force.two}{logical to use for counts where values >1 should be considered 1 (then effectively testing and showing a binary variable).}

\item{show.na}{logical to use missing data as an explicit category (default FALSE)}
}
\value{
returns a data.frame. Has standardized column names, ready to stack with output from other grit::twoway* functions.
}
\description{
twoway table of categorical data
}
