% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_num.R
\name{twoway_num}
\alias{twoway_num}
\title{twoway table of numeric data}
\usage{
twoway_num(data, x, group, digit.m = 1, digit.sd = 1, cal.date = F)
}
\arguments{
\item{data}{data frame}

\item{x}{the numeric variable to be summarized}

\item{group}{the variable which identifies treatment/comparison groups}

\item{digit.m}{number of digits on mean estimates}

\item{digit.sd}{number of digits for estimates of the standard deviation}

\item{cal.date}{logical indicating that the x variable should be treated as a date. Will then show the mean value as an actual date.}
}
\description{
twoway table of numeric data
}
